refactor: renderTiptapEditorで統一
