+++ src/features/editor/extensions/nodes/code-block-container/code-block/__tests__/command.spec.ts
@@ -5,7 +5,7 @@ import { ListKit } from "@tiptap/extension-list";
 import Paragraph from "@tiptap/extension-paragraph";
 import Text from "@tiptap/extension-text";
 import { describe, expect, it } from "vitest";
-import { createEditorInstance } from "@/tests/editor-instance";
+import { renderTiptapEditor } from "@/tests/editor";
 import { Details } from "../../../details";
 import { DetailsContent } from "../../../details/content";
 import { DetailsSummary } from "../../../details/summary";
@@ -31,7 +31,7 @@ const baseExtensions = [
 describe("コマンド", () => {
   describe("setAllSelectionInCodeBlock", () => {
     it("コードブロック全体を選択できる", () => {
-      const editor = createEditorInstance({
+      const editor = renderTiptapEditor({
         extensions: baseExtensions,
         content:
           '<div class="code-block-container"><div class="code-block-filename-container"><span class="code-block-filename"></span></div><pre><code class="language-javascript">console.log("hello");</code></pre></div>',
@@ -52,7 +52,7 @@ describe("コマンド", () => {
 
   describe("setCodeBlockContainer", () => {
     it("段落をコードブロックに変換できる", () => {
-      const editor = createEditorInstance({
+      const editor = renderTiptapEditor({
         extensions: baseExtensions,
         content: '<p>console.log("hello");</p>',
       });
@@ -70,7 +70,7 @@ describe("コマンド", () => {
     });
 
     it("改行ありの段落を保持したまま呼び出せる", () => {
-      const editor = createEditorInstance({
+      const editor = renderTiptapEditor({
         extensions: baseExtensions,
         content: '<p>console.log("hello");<br>console.log("world");</p>',
       });
@@ -88,7 +88,7 @@ describe("コマンド", () => {
     });
 
     it("見出しの中で呼び出せる", () => {
-      const editor = createEditorInstance({
+      const editor = renderTiptapEditor({
         extensions: [...baseExtensions, Heading],
         content: "<h1>見出しの中</h1>",
       });
@@ -102,7 +102,7 @@ describe("コマンド", () => {
     });
 
     it("引用の中で呼び出せる", () => {
-      const editor = createEditorInstance({
+      const editor = renderTiptapEditor({
         extensions: [...baseExtensions, Blockquote],
         content: "<blockquote>引用の中</blockquote>",
       });
@@ -116,7 +116,7 @@ describe("コマンド", () => {
     });
 
     it("リストで呼び出せない", () => {
-      const editor = createEditorInstance({
+      const editor = renderTiptapEditor({
         extensions: [...baseExtensions, ListKit],
         content: "<ul><li><p>リストの中</p></li></ul>",
       });
@@ -130,7 +130,7 @@ describe("コマンド", () => {
     });
 
     it("メッセージの中で呼び出せる", () => {
-      const editor = createEditorInstance({
+      const editor = renderTiptapEditor({
         extensions: baseExtensions,
         content:
           '<aside class="msg"><div class="msg-content"><p>メッセージ</p></div></aside>',
@@ -145,7 +145,7 @@ describe("コマンド", () => {
     });
 
     it("コードブロックのファイル名とコンテンツの中で呼び出せない", () => {
-      const editor = createEditorInstance({
+      const editor = renderTiptapEditor({
         extensions: [
           ...baseExtensions,
           CodeBlock,
@@ -173,7 +173,7 @@ describe("コマンド", () => {
     });
 
     it("差分コードブロックのファイル名とコンテンツの中で呼び出せない", () => {
-      const editor = createEditorInstance({
+      const editor = renderTiptapEditor({
         extensions: [
           ...baseExtensions,
           CodeBlock,
@@ -201,7 +201,7 @@ describe("コマンド", () => {
     });
 
     it("アコーディオンのサマリーで呼べない。コンテンツで呼べる", () => {
-      const editor = createEditorInstance({
+      const editor = renderTiptapEditor({
         extensions: [
           ...baseExtensions,
           Details,
@@ -229,7 +229,7 @@ describe("コマンド", () => {
 
   describe("unsetCodeBlockContainer", () => {
     it("コードブロックを段落に戻せる", () => {
-      const editor = createEditorInstance({
+      const editor = renderTiptapEditor({
         extensions: baseExtensions,
         content:
           '<div class="code-block-container"><div class="code-block-filename-container"><span class="code-block-filename"></span></div><pre><code class="language-javascript">console.log("hello");</code></pre></div>',
