+++ src/features/editor/extensions/nodes/code-block-container/code-block/__tests__/markdown.spec.ts
@@ -5,7 +5,7 @@ import Text from "@tiptap/extension-text";
 import { describe, expect, it } from "vitest";
 import { fromMarkdown } from "@/features/editor/lib/from-markdown";
 import { markdownSerializer } from "@/features/editor/lib/to-markdown";
-import { createEditorInstance } from "@/tests/editor-instance";
+import { renderTiptapEditor } from "@/tests/editor";
 import { CodeBlockFileName } from "../../code-block-file-name";
 import { DiffCodeBlock } from "../../diff-code-block";
 import { DiffCodeLine } from "../../diff-code-block/diff-code-line";
@@ -26,7 +26,7 @@ const basicExtension = [
 
 describe("マークダウン", () => {
   it("JavaScriptコードブロックをマークダウンに変換できる", () => {
-    const editor = createEditorInstance({
+    const editor = renderTiptapEditor({
       extensions: basicExtension,
       content:
         '<div class="code-block-container"><div class="code-block-filename-container"><span class="code-block-filename"></span></div><pre><code class="language-javascript">console.log("hello");</code></pre></div>',
@@ -37,7 +37,7 @@ describe("マークダウン", () => {
   });
 
   it("言語名なしのコードブロックをマークダウンに変換できる", () => {
-    const editor = createEditorInstance({
+    const editor = renderTiptapEditor({
       extensions: basicExtension,
       content:
         '<div class="code-block-container"><div class="code-block-filename-container"><span class="code-block-filename"></span></div><pre><code class="language-plaintext">plaintext code</code></pre></div>',
@@ -48,7 +48,7 @@ describe("マークダウン", () => {
   });
 
   it("複数行のコードブロックをマークダウンに変換できる", () => {
-    const editor = createEditorInstance({
+    const editor = renderTiptapEditor({
       extensions: basicExtension,
       content:
         '<div class="code-block-container"><div class="code-block-filename-container"><span class="code-block-filename"></span></div><pre><code class="language-python">def hello():\n    print("Hello, World!")\n    return True</code></pre></div>',
@@ -64,7 +64,7 @@ describe("マークダウン", () => {
     const markdown = '```javascript\nconsole.log("hello");\n```';
 
     const html = fromMarkdown(markdown);
-    const editor = createEditorInstance({
+    const editor = renderTiptapEditor({
       extensions: basicExtension,
       content: html,
     });
@@ -79,7 +79,7 @@ describe("マークダウン", () => {
     const markdown = '```javascript:hello.js\nconsole.log("hello");\n```';
 
     const html = fromMarkdown(markdown);
-    const editor = createEditorInstance({
+    const editor = renderTiptapEditor({
       extensions: basicExtension,
       content: html,
     });
