+++ src/features/editor/extensions/nodes/code-block-container/code-block/__tests__/render.spec.ts
@@ -3,7 +3,7 @@ import HardBreak from "@tiptap/extension-hard-break";
 import Paragraph from "@tiptap/extension-paragraph";
 import Text from "@tiptap/extension-text";
 import { describe, expect, it } from "vitest";
-import { createEditorInstance } from "@/tests/editor-instance";
+import { renderTiptapEditor } from "@/tests/editor";
 import { CodeBlockFileName } from "../../code-block-file-name";
 import { DiffCodeBlock } from "../../diff-code-block";
 import { DiffCodeLine } from "../../diff-code-block/diff-code-line";
@@ -26,7 +26,7 @@ describe("HTMLのレンダリング", () => {
   it("JavaScriptコードブロックが正しいHTMLでレンダリングされる", () => {
     const content =
       '<div class="code-block-container"><div class="code-block-filename-container"><span class="code-block-filename"></span></div><pre><code class="language-javascript">console.log("hello");</code></pre></div>';
-    const editor = createEditorInstance({
+    const editor = renderTiptapEditor({
       extensions: basicExtension,
       content,
     });
@@ -38,7 +38,7 @@ describe("HTMLのレンダリング", () => {
   it("言語名なしのコードブロックが正しいHTMLでレンダリングされる", () => {
     const content =
       '<div class="code-block-container"><div class="code-block-filename-container"><span class="code-block-filename"></span></div><pre><code class="language-plaintext">plaintext code</code></pre></div>';
-    const editor = createEditorInstance({
+    const editor = renderTiptapEditor({
       extensions: basicExtension,
       content,
     });
