+++ src/features/editor/extensions/nodes/code-block-container/diff-code-block/__tests__/command.spec.ts
@@ -3,7 +3,7 @@ import HardBreak from "@tiptap/extension-hard-break";
 import Paragraph from "@tiptap/extension-paragraph";
 import Text from "@tiptap/extension-text";
 import { describe, expect, it } from "vitest";
-import { createEditorInstance } from "@/tests/editor-instance";
+import { renderTiptapEditor } from "@/tests/editor";
 
 import { CodeBlock } from "../../code-block";
 import { CodeBlockFileName } from "../../code-block-file-name";
@@ -27,7 +27,7 @@ const baseExtensions = [
 describe("コマンド", () => {
   describe("setAllSelectionInCodeBlock", () => {
     it("1行の差分コードブロック全体を選択できる", () => {
-      const editor = createEditorInstance({
+      const editor = renderTiptapEditor({
         extensions: baseExtensions,
         content:
           '<div class="code-block-container"><div class="code-block-filename-container"><span class="code-block-filename"></span></div><pre><code class="language-diff-javascript diff-highlight"><span>console.log("hello");</span></code></pre></div>',
@@ -46,7 +46,7 @@ describe("コマンド", () => {
     });
 
     it("複数行の差分コードブロック全体を選択できる", () => {
-      const editor = createEditorInstance({
+      const editor = renderTiptapEditor({
         extensions: baseExtensions,
         content: `<div class="code-block-container"><div class="code-block-filename-container"><span class="code-block-filename"></span></div>
         <pre><code class="language-diff diff-highlight"><span>1</span><span>2</span><span>3</span></code></pre></div>`,
@@ -67,7 +67,7 @@ describe("コマンド", () => {
 
   describe("setCodeBlockContainer", () => {
     it("段落を差分コードブロックに変換できる", () => {
-      const editor = createEditorInstance({
+      const editor = renderTiptapEditor({
         extensions: baseExtensions,
         content: '<p>console.log("hello");</p>',
       });
@@ -85,7 +85,7 @@ describe("コマンド", () => {
     });
 
     it("改行ありの段落を保持したまま変換できる", () => {
-      const editor = createEditorInstance({
+      const editor = renderTiptapEditor({
         extensions: baseExtensions,
         content: "<p>const a = 1;<br><br>const a = 2;</p>",
       });
@@ -103,7 +103,7 @@ describe("コマンド", () => {
     });
 
     it("段落を跨ぐ範囲選択時に変換できる", () => {
-      const editor = createEditorInstance({
+      const editor = renderTiptapEditor({
         extensions: baseExtensions,
         content: "<p>const a = 1;</p><p>const a = 2;</p>",
       });
@@ -123,7 +123,7 @@ describe("コマンド", () => {
 
   describe("unsetCodeBlockContainer", () => {
     it("差分コードブロックを段落に戻せる", () => {
-      const editor = createEditorInstance({
+      const editor = renderTiptapEditor({
         extensions: baseExtensions,
         content: `<div class="code-block-container"><div class="code-block-filename-container"><span class="code-block-filename"></span></div>
         <pre><code class="language-diff-javascript diff-highlight"><span>const a= 1;</span></code></pre></div>`,
@@ -140,7 +140,7 @@ describe("コマンド", () => {
     });
 
     it("複数行持つ差分コードブロックを段落に戻せる", () => {
-      const editor = createEditorInstance({
+      const editor = renderTiptapEditor({
         extensions: baseExtensions,
         content: `<div class="code-block-container"><div class="code-block-filename-container"><span class="code-block-filename"></span></div>
         <pre><code class="language-diff-javascript diff-highlight"><span>const a = 1;</span><span>const a = 2;</span></code></pre></div>`,
