+++ src/features/editor/extensions/nodes/code-block-container/diff-code-block/__tests__/parse.spec.ts
@@ -3,7 +3,7 @@ import HardBreak from "@tiptap/extension-hard-break";
 import Paragraph from "@tiptap/extension-paragraph";
 import Text from "@tiptap/extension-text";
 import { describe, expect, it } from "vitest";
-import { createEditorInstance } from "@/tests/editor-instance";
+import { renderTiptapEditor } from "@/tests/editor";
 import { CodeBlock } from "../../code-block";
 import { CodeBlockFileName } from "../../code-block-file-name";
 import { CodeBlockContainer } from "../../index";
@@ -24,7 +24,7 @@ const baseExtensions = [
 
 describe("HTMLのパース", () => {
   it("preタグを差分コードブロックノードとしてパースできる", () => {
-    const editor = createEditorInstance({
+    const editor = renderTiptapEditor({
       extensions: baseExtensions,
       content: `<div class="code-block-container"><div class="code-block-filename-container"><span class="code-block-filename"></span></div>
       <pre><code class="language-diff-javascript diff-highlight"><span>console.log("hello");</span></code></pre></div>`,
@@ -42,7 +42,7 @@ describe("HTMLのパース", () => {
   });
 
   it("言語名とファイル名ありの差分preタグをパースできる", () => {
-    const editor = createEditorInstance({
+    const editor = renderTiptapEditor({
       extensions: baseExtensions,
       content: `<div class="code-block-container"><div class="code-block-filename-container"><span class="code-block-filename">example.ts</span></div>
       <pre><code class="language-diff-python diff-highlight"><span>import os</span></code></pre></div>`,
@@ -60,7 +60,7 @@ describe("HTMLのパース", () => {
   });
 
   it("diff言語をパースできる", () => {
-    const editor = createEditorInstance({
+    const editor = renderTiptapEditor({
       extensions: baseExtensions,
       content: `<div class="code-block-container"><div class="code-block-filename-container"><span class="code-block-filename">example.ts</span></div>
       <pre><code class="language-diff diff-highlight"><span>const a = 1;</span></code></pre></div>`,
@@ -78,7 +78,7 @@ describe("HTMLのパース", () => {
   });
 
   it("複数行でもパースできる", () => {
-    const editor = createEditorInstance({
+    const editor = renderTiptapEditor({
       extensions: baseExtensions,
       content: `<div class="code-block-container"><div class="code-block-filename-container"><span class="code-block-filename">example.ts</span></div>
       <pre><code class="language-diff diff-highlight"><span>const a = 1;</span><span>const b = 2;</span></code></pre></div>`,
