+++ src/features/editor/extensions/nodes/code-block-container/diff-code-block/__tests__/render.spec.ts
@@ -3,7 +3,7 @@ import HardBreak from "@tiptap/extension-hard-break";
 import Paragraph from "@tiptap/extension-paragraph";
 import Text from "@tiptap/extension-text";
 import { describe, expect, it } from "vitest";
-import { createEditorInstance } from "@/tests/editor-instance";
+import { renderTiptapEditor } from "@/tests/editor";
 import { CodeBlockContainer } from "../..";
 import { CodeBlock } from "../../code-block";
 import { CodeBlockFileName } from "../../code-block-file-name";
@@ -26,7 +26,7 @@ describe("HTMLのレンダリング", () => {
   it("JavaScript差分コードブロックが正しいHTMLでレンダリングされる", () => {
     const content =
       '<div class="code-block-container"><div class="code-block-filename-container"><span class="code-block-filename"></span></div><pre><code class="diff-highlight language-diff-javascript"><span>console.log("hello");</span></code></pre></div>';
-    const editor = createEditorInstance({
+    const editor = renderTiptapEditor({
       extensions: baseExtensions,
       content,
     });
@@ -38,7 +38,7 @@ describe("HTMLのレンダリング", () => {
   it("diff言語の差分コードブロックが正しいHTMLでレンダリングされる", () => {
     const content =
       '<div class="code-block-container"><div class="code-block-filename-container"><span class="code-block-filename"></span></div><pre><code class="diff-highlight language-diff"><span>plaintext code</span></code></pre></div>';
-    const editor = createEditorInstance({
+    const editor = renderTiptapEditor({
       extensions: baseExtensions,
       content,
     });
