+++ src/features/editor/extensions/nodes/details/__tests__/parse.spec.ts
@@ -2,7 +2,7 @@ import Document from "@tiptap/extension-document";
 import Paragraph from "@tiptap/extension-paragraph";
 import Text from "@tiptap/extension-text";
 import { describe, expect, it } from "vitest";
-import { createEditorInstance } from "@/tests/editor-instance";
+import { renderTiptapEditor } from "@/tests/editor";
 import { Details } from "..";
 import { DetailsContent } from "../content";
 import { DetailsSummary } from "../summary";
@@ -18,7 +18,7 @@ const basicExtension = [
 
 describe("HTMLのパース・レンダリング", () => {
   it("アコーディオンのHTMLをパースできる", () => {
-    const editor = createEditorInstance({
+    const editor = renderTiptapEditor({
       extensions: basicExtension,
       content:
         '<details><summary></summary><div class="details-content"><p>テキスト</p></div></details>',
