+++ src/features/editor/extensions/nodes/figure/__tests__/paste.browser.spec.ts
@@ -4,8 +4,9 @@ import Text from "@tiptap/extension-text";
 import { userEvent } from "@vitest/browser/context";
 import { describe, expect, it } from "vitest";
 import LakeImage from "@/assets/sikotuko.jpeg";
+import { copyText } from "@/tests/clipboard";
 import { waitSelectionChange } from "@/tests/dom";
-import { copyText, renderTiptapEditor } from "@/tests/editor";
+import { renderTiptapEditor } from "@/tests/editor";
 import { Details } from "../../details";
 import { DetailsContent } from "../../details/content";
 import { DetailsSummary } from "../../details/summary";
