+++ src/features/editor/extensions/nodes/figure/__tests__/render.spec.ts
@@ -3,7 +3,7 @@ import Paragraph from "@tiptap/extension-paragraph";
 import Text from "@tiptap/extension-text";
 import { describe, expect, it } from "vitest";
 import LakeImage from "@/assets/sikotuko.jpeg";
-import { createEditorInstance } from "@/tests/editor-instance";
+import { renderTiptapEditor } from "@/tests/editor";
 import Figure from "..";
 import { Caption } from "../caption";
 import { Image } from "../image";
@@ -12,7 +12,7 @@ const basicExtension = [Document, Paragraph, Text, Figure, Image, Caption];
 
 describe("HTMLのパース・レンダリング", () => {
   it("Figureノードが正しいHTMLでレンダリングされる", () => {
-    const editor = createEditorInstance({
+    const editor = renderTiptapEditor({
       extensions: basicExtension,
       content: `<p><img src="${LakeImage}" alt="支笏湖"><em>支笏湖</em></p>`,
     });
