+++ src/features/editor/extensions/nodes/message/__tests__/parser.spec.ts
@@ -2,7 +2,7 @@ import Document from "@tiptap/extension-document";
 import Paragraph from "@tiptap/extension-paragraph";
 import Text from "@tiptap/extension-text";
 import { describe, expect, it } from "vitest";
-import { createEditorInstance } from "@/tests/editor-instance";
+import { renderTiptapEditor } from "@/tests/editor";
 import { Message } from "../message";
 import { MessageContent } from "../message-content";
 
@@ -10,7 +10,7 @@ const basicExtension = [Document, Paragraph, Text, Message, MessageContent];
 
 describe("HTMLのパース", () => {
   it("aside.msgをメッセージノードとしてパースできる", () => {
-    const editor = createEditorInstance({
+    const editor = renderTiptapEditor({
       extensions: basicExtension,
       content:
         '<aside class="msg"><div class="msg-content"><p>メッセージ</p></div></aside>',
@@ -24,7 +24,7 @@ describe("HTMLのパース", () => {
   });
 
   it("aside.msg.alertをアラートタイプとしてパースできる", () => {
-    const editor = createEditorInstance({
+    const editor = renderTiptapEditor({
       extensions: basicExtension,
       content:
         '<aside class="msg alert"><div class="msg-content"><p>メッセージ</p></div></aside>',
