+++ src/features/editor/extensions/nodes/message/__tests__/render.spec.ts
@@ -2,7 +2,7 @@ import Document from "@tiptap/extension-document";
 import Paragraph from "@tiptap/extension-paragraph";
 import Text from "@tiptap/extension-text";
 import { describe, expect, it } from "vitest";
-import { createEditorInstance } from "@/tests/editor-instance";
+import { renderTiptapEditor } from "@/tests/editor";
 import { Message } from "../message";
 import { MessageContent } from "../message-content";
 
@@ -10,7 +10,7 @@ const basicExtension = [Document, Paragraph, Text, Message, MessageContent];
 
 describe("HTMLのレンダリング", () => {
   it("メッセージタイプが正しいHTMLでレンダリングされる", () => {
-    const editor = createEditorInstance({
+    const editor = renderTiptapEditor({
       extensions: basicExtension,
       content:
         '<aside class="msg"><div class="msg-content"><p>メッセージ</p></div></aside>',
@@ -22,7 +22,7 @@ describe("HTMLのレンダリング", () => {
   });
 
   it("アラートタイプが正しいHTMLでレンダリングされる", () => {
-    const editor = createEditorInstance({
+    const editor = renderTiptapEditor({
       extensions: basicExtension,
       content:
         '<aside class="msg alert"><div class="msg-content"><p>メッセージ</p></div></aside>',
@@ -33,7 +33,7 @@ describe("HTMLのレンダリング", () => {
   });
 
   it("HTMLにレンダリング後にパースできる", () => {
-    const editor = createEditorInstance({
+    const editor = renderTiptapEditor({
       extensions: basicExtension,
       content:
         '<aside class="msg alert"><div class="msg-content"><p>メッセージ</p></div></aside>',
